# app/services/tuber-api/v1/docs/swagger.json
{
    "schemes": [
        "http",
        "https",
        "ws",
        "wss"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the TUber application.",
        "title": "TUber API",
        "contact": {},
        "version": "develop"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "paths": {
...
        "/trips": {
            "get": {
                "description": "Query will query trips",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trip"
                ],
                "summary": "get all trips",
                "responses": {
                    "200": {
                        "description": "query trips",
                        "schema": {
                            "$ref": "#/definitions/tripgrp.AppTrip"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Create will add a trip",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trip"
                ],
                "summary": "create a new trip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "New Trip",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tripgrp.AppNewTrip"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Trip successfully created",
                        "schema": {
                            "$ref": "#/definitions/tripgrp.AppTrip"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/trips/my": {
            "get": {
                "description": "QueryMyTrip will query trips by user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trip"
                ],
                "summary": "get my trips",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Is Driver",
                        "name": "is_driver",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "query trips",
                        "schema": {
                            "$ref": "#/definitions/tripgrp.AppTrip"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/trips/{id}": {
            "get": {
                "description": "QueryByID will query trips by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trip"
                ],
                "summary": "get trips by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "query trips",
                        "schema": {
                            "$ref": "#/definitions/tripgrp.AppTrip"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "Update will update a trip",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trip"
                ],
                "summary": "update a trip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Trip",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tripgrp.AppUpdateTrip"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip successfully updated",
                        "schema": {
                            "$ref": "#/definitions/tripgrp.AppTrip"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/trips/{id}/join": {
            "post": {
                "description": "Join will join a trip",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trip"
                ],
                "summary": "join a trip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip successfully joined",
                        "schema": {
                            "$ref": "#/definitions/tripgrp.AppTrip"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/trips/{id}/passengers": {
            "get": {
                "description": "QueryPassengers will query passengers of a trip",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trip"
                ],
                "summary": "get all passengers of a trip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "query passengers of a trip",
                        "schema": {
                            "$ref": "#/definitions/tripgrp.AppTripDetails"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/trips/{id}/rating": {
            "post": {
                "description": "Create will add a rating",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trip"
                ],
                "summary": "create a new rating for a trip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "New Rating",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tripgrp.AppNewRating"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Rating successfully created",
                        "schema": {
                            "$ref": "#/definitions/tripgrp.AppRating"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "tripgrp.AppNewRating": {
            "type": "object",
            "required": [
                "rating"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                }
            }
        },
        "tripgrp.AppNewTrip": {
            "type": "object",
            "required": [
                "start_time"
            ],
            "properties": {
                "destination": {
                    "$ref": "#/definitions/tripgrp.AppNewTripLocation"
                },
                "mid": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tripgrp.AppNewTripLocation"
                    }
                },
                "passenger_limit": {
                    "type": "integer"
                },
                "source": {
                    "$ref": "#/definitions/tripgrp.AppNewTripLocation"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "tripgrp.AppNewTripLocation": {
            "type": "object",
            "required": [
                "lat",
                "lon",
                "name",
                "place_id"
            ],
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "place_id": {
                    "type": "string"
                }
            }
        },
        "tripgrp.AppPassengerDetails": {
            "type": "object",
            "properties": {
                "destination_latitude": {
                    "type": "number"
                },
                "destination_longitude": {
                    "type": "number"
                },
                "destination_name": {
                    "type": "string"
                },
                "destination_place_id": {
                    "type": "string"
                },
                "passenger_id": {
                    "type": "string"
                },
                "source_latitude": {
                    "type": "number"
                },
                "source_longitude": {
                    "type": "number"
                },
                "source_name": {
                    "type": "string"
                },
                "source_place_id": {
                    "type": "string"
                }
            }
        },
        "tripgrp.AppRating": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "commenter_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "trip_id": {
                    "type": "string"
                }
            }
        },
        "tripgrp.AppTrip": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "destination_id": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "passenger_limit": {
                    "type": "integer"
                },
                "source_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "tripgrp.AppTripDetails": {
            "type": "object",
            "properties": {
                "destination_latitude": {
                    "type": "number"
                },
                "destination_longitude": {
                    "type": "number"
                },
                "destination_name": {
                    "type": "string"
                },
                "destination_place_id": {
                    "type": "string"
                },
                "driver_details": {
                    "$ref": "#/definitions/tripgrp.AppTripDriverDetails"
                },
                "passenger_details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tripgrp.AppPassengerDetails"
                    }
                },
                "source_latitude": {
                    "type": "number"
                },
                "source_longitude": {
                    "type": "number"
                },
                "source_name": {
                    "type": "string"
                },
                "source_place_id": {
                    "type": "string"
                },
                "trip_id": {
                    "type": "string"
                }
            }
        },
        "tripgrp.AppTripDriverDetails": {
            "type": "object",
            "properties": {
                "driver_brand": {
                    "type": "string"
                },
                "driver_color": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "string"
                },
                "driver_image_url": {
                    "type": "string"
                },
                "driver_model": {
                    "type": "string"
                },
                "driver_name": {
                    "type": "string"
                },
                "driver_plate": {
                    "type": "string"
                }
            }
        },
        "tripgrp.AppUpdateTrip": {
            "type": "object",
            "properties": {
                "passenger_limit": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}